44. Admin Orders

+++ frontend/src/helpers/axios.js
@@ -25,9 +25,9 @@ axiosIntance.interceptors.response.use(
     return res;
   },
   (error) => {
-    console.log({ error });
-    const { status } = error.response;
-    if (status === 500) {
+    console.log(error.response);
+    const { status } = error.response ? error.response.status : 500;
+    if (status && status === 500) {
       localStorage.clear();
       store.dispatch({ type: authConstants.LOGOUT_SUCCESS });
     }
